# Reduce small string memory overhead by ~20-30%