@madolson 2025-08-20
We probably also want a performance test for IO threading as well either that or a deep pipeline. It will tease out a performance regression more clearly.