@rainsupreme 2026-01-06
I've rebased again. This time I've omitted the `pure` markers because it did not improve performance.

My conclusions from the performance tests:
- The added complexity in determining the location of the embedded value data caused a ~4% performance decrease for SET performance, and no clear change to GET performance. I'd argue losing ~4% SET perf to eliminate ~25% of memory overhead cost is worth it. (Note: after this change it would be pretty easy to maintain a fork with different behavior or change it here. It's easy to undo this tradeoff. 😉)
- There are different performance patterns for embedded vs non-embedded values. The changes to the 64B value size reflect the change from non-embedded to embedded caused by raising the size threshold.
- For some reason when compiled with -O2, SET commands perform up to ~70% faster. I am surprised, curious, excited, and I plan to investigate this further in the future. Feel free to ping me if you want to chat about that finding 😀