@zuiderkwast 2026-01-06
Great, finally!

> After discussion with @madolson I've reverted the increase in size threshold for embedding values for now

So the memory usage is the green line in the "Memory efficiency comparison" diagram. Makes sense.

For a 16B key, it means we only embed values up to 33 bytes, instead of up to 97 bytes.

    +--------+-----------------+------------+------------+
    | header | key header size | key sds5   | val sds8   |
    | 8      | 1               | 1 + 16 + 1 | 3 + 33 + 1 |  = 64 bytes
    +--------+-----------------+------------+------------+

    +--------+-----------------+------------+------------+
    | header | key header size | key sds5   | val sds8   |
    | 8      | 1               | 1 + 16 + 1 | 3 + 97 + 1 |  = 128 bytes
    +--------+-----------------+------------+------------+

Many string values fall in the range 34-97 bytes, so it a little sad that we can't apply this memory saving to these values. Hopefully we can solve the performence issue in the future and increase the threshold later.